requirements.txt
aiogram>=3.13,<4
google-api-python-client>=2.150,<3
google-auth>=2.35,<3
pyairtable>=2.3,<3
google-genai>=1.0,<2
pydantic>=2.10,<3
python-dotenv>=1.0,<2
requests>=2.32,<3
